post repository test complete

+++ backEnd/src/main/java/org/movieproject/posts/repository/PostsRepository.java
@@ -4,6 +4,6 @@ import org.movieproject.posts.dto.PostsDTO;
 import org.movieproject.posts.entity.Posts;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostsRepository extends JpaRepository<Posts, Integer> {
+public interface PostsRepository extends JpaRepository<Posts, Long> {
 
 }
